+++ R/ma_support.R
@@ -93,10 +93,7 @@ ma_article_aside <- function(x){
   cols <- gsub(":$", "", cols)
   cols <- gsub("\\s", "_", cols)
   x <- rvest::html_nodes(x, ".pi-data-value")
-  if(length(x) != length(cols)){
-    warning("Summary card cannot be parsed. Metadata column will be NULL.")
-    return()
-  }
+  if(length(x) != length(cols)) return()
   vals <- purrr::map(x, ~{
     x <- xml2::xml_contents(.x) %>% ma_text(trim = FALSE)
     x[x %in% c("", " ")] <- "|"